+++ src/Farkle/Exceptions.fs
@@ -7,20 +7,24 @@ namespace Farkle
 
 open Farkle.Grammar
 open Farkle.Parser
+open System.Runtime.CompilerServices
+open System.Runtime.InteropServices
 
 /// The base class for all exceptions
 /// that might be thrown from Farkle.
-type FarkleException(msg: string, innerExn) =
+type FarkleException([<Optional; DefaultParameterValue(null: string); Nullable(2uy)>] msg: string,
+    [<Optional; Nullable(2uy)>] innerExn) =
     inherit exn(msg, innerExn)
-    new msg = FarkleException(msg, null)
 
 /// An exception thrown by Farkle's parser on a snytax or lexical error.
 type ParserException(error: ParserError) =
-    inherit FarkleException(error.ToString(), null)
+    inherit FarkleException(error.ToString())
     /// <summary>The <see cref="ParserError"/>
     /// object this exception holds.</summary>
     member _.Error = error
 
+/// An exception thrown by Farkle when the post-processor throws
+/// another exception. That exception is wrapped to this one.
 type PostProcessorException private(msg, innerExn) =
     inherit FarkleException(msg, innerExn)
     new (term: Terminal, innerExn) =
@@ -36,4 +40,4 @@ type PostProcessorException private(msg, innerExn) =
 /// will be caught and rewrapped as <see cref="ParserException"/>s.
 /// F# users can use the <c>error</c> or <c>errorf</c> functions
 /// in <c>Farkle.Builder</c>.</remarks>
-type ParserApplicationException(msg) = inherit FarkleException(msg, null)
+type ParserApplicationException(msg) = inherit FarkleException(msg)